Refactor codebase for cross-platform compatibility and improved memory management
This commit includes a major refactoring of the codebase. The changes are aimed at improving cross-platform compatibility, memory management, and database interaction.

The include paths have been updated to use forward slashes for cross-platform compatibility. The `AgentData` and `FinancialData` classes have been significantly modified, with changes to constructors, member variables, and methods. The `AgentData` class no longer interacts directly with the database, and the `FinancialData` class now stores training and test data records together in a single vector.

The `Database::SQLCommand` and `Database::SQLConnection` classes have been expanded and refactored, with changes to constructors, destructors, and methods. The code has been switched from using ODBC to OLE DB for connecting to the SQL Server database.

The `pushGP::globals::population_agents` array has been changed to a pointer, suggesting dynamic allocation. The `make_pop_agents` function now interacts with a `Plush::Environment` object.

The `Run.DevelopStrategy_V2.cpp` and `Run.DevelopStrategy_V2.h` files have been stripped of all content, and the `Arguments.h` file has been updated with new constants.

The code has shifted from using raw pointers to manage dynamic arrays of `Individual` objects to using `std::unique_ptr` with `std::array`. This change provides better memory management and safety.

The `EventLogManager.cpp` and `EventLogManager.h` files have been completely removed, suggesting a change in the logging functionality.
