+++ SOS/Database/SQLConnection.cpp
@@ -2,183 +2,404 @@
 #include <iostream>
 #include <thread>
 #include "SQLConnection.h"
+#include "..\Utilities\Conversion.h"
+//#include "..\Utilities\EventLogManager.h"
+#include "..\Utilities\Debug.h"
+#include "..\Utilities\Conversion.h"
 
-namespace Database
+namespace database
 {
-	//SQLConnection::SQLConnection()
+	//HRESULT SQLConnection::initialize_and_establish_connection(const OLECHAR * server, const OLECHAR * db_string, const OLECHAR * user_id, const OLECHAR * password)
 	//{
-	//	hstmt1 = SQL_NULL_HSTMT;
-	//	hdbc1 = SQL_NULL_HDBC;
-	//	henv = SQL_NULL_HENV;
+	//	DBPROP          InitProperties[5] = { 0 };
+	//	DBPROPSET       rgInitPropSet[1] = { 0 };
+	//	HRESULT         hr = S_OK;
+
+	//	try
+	//	{
+	//		int attempts = 0;
+
+	//		do
+	//		{
+	//			attempts++;
+
+	//			// Obtain access to the OLE DB Driver for SQL Server.  
+	//			hr = CoCreateInstance(MSOLEDBSQL_CLSID,
+	//				NULL,
+	//				CLSCTX_INPROC_SERVER,
+	//				IID_IDBInitialize,
+	//				(void **)&pIDBInitialize_);
+	//			
+	//			if (FAILED(hr))
+	//			{
+	//				char error_message[80];
+	//				sprintf_s(error_message, 80, "Attempt %d.  Failed to obtain access to the OLE DB Driver.  hr = %lX", attempts, hr);
+
+	//				std::cout << error_message << std::endl;
+
+	//				std::this_thread::sleep_for(std::chrono::milliseconds(500));
+
+	//				if (attempts == 3)
+	//					throw std::runtime_error(error_message);
+	//			}
+	//		}
+	//		while (FAILED(hr));
+
+	//		// Initialize property values needed to establish connection.  
+	//		for (int i = 0; i < 5; i++)
+	//			VariantInit(&InitProperties[i].vValue);
+
+	//		// Server name.  
+	//		// See DBPROP structure for more information on InitProperties  
+	//		InitProperties[0].dwPropertyID = DBPROP_INIT_DATASOURCE;
+	//		InitProperties[0].vValue.vt = VT_BSTR;
+	//		InitProperties[0].vValue.bstrVal = SysAllocString(server);
+	//		InitProperties[0].dwOptions = DBPROPOPTIONS_REQUIRED;
+	//		InitProperties[0].colid = DB_NULLID;
+
+	//		OLECHAR c = '\0';
+
+	//		// Database.  
+	//		InitProperties[1].dwPropertyID = DBPROP_INIT_CATALOG;
+	//		InitProperties[1].vValue.vt = VT_BSTR;
+	//		InitProperties[1].vValue.bstrVal = SysAllocString(db_string);
+	//		InitProperties[1].dwOptions = DBPROPOPTIONS_REQUIRED;
+	//		InitProperties[1].colid = DB_NULLID;
+
+	//		// Username (login).  
+	//		InitProperties[2].dwPropertyID = DBPROP_AUTH_USERID;
+	//		InitProperties[2].vValue.vt = VT_BSTR;
+	//		InitProperties[2].vValue.bstrVal = SysAllocString(user_id);
+	//		InitProperties[2].dwOptions = DBPROPOPTIONS_REQUIRED;
+	//		InitProperties[2].colid = DB_NULLID;
+
+	//		// Password (login)
+	//		InitProperties[3].dwPropertyID = DBPROP_AUTH_PASSWORD;
+	//		InitProperties[3].vValue.vt = VT_BSTR;
+	//		InitProperties[3].vValue.bstrVal = SysAllocString(password);
+	//		InitProperties[3].dwOptions = DBPROPOPTIONS_REQUIRED;
+	//		InitProperties[3].colid = DB_NULLID;
+
+	//		// Encrypt=Optional
+	//		const OLECHAR* encryption = L"Optional";
+	//		InitProperties[4].dwPropertyID = SSPROP_INIT_ENCRYPT;
+	//		InitProperties[4].vValue.vt = VT_BSTR;
+	//		InitProperties[4].vValue.bstrVal = SysAllocString(encryption);
+	//		InitProperties[4].dwOptions = DBPROPOPTIONS_REQUIRED;
+	//		InitProperties[4].colid = DB_NULLID;
+
+	//		// Construct the DBPROPSET structure(rgInitPropSet). The   
+	//		// DBPROPSET structure is used to pass an array of DBPROP   
+	//		// structures (InitProperties) to the SetProperties method.  
+	//		rgInitPropSet[0].guidPropertySet = DBPROPSET_DBINIT;
+	//		rgInitPropSet[0].cProperties = 5;
+	//		rgInitPropSet[0].rgProperties = InitProperties;
+
+	//		// Set initialization properties.  
+	//		hr = pIDBInitialize_->QueryInterface(IID_IDBProperties, (void **)&pIDBProperties_);
+	//		if (FAILED(hr))
+	//			throw std::runtime_error("Failed to obtain an IDBProperties interface.");
+
+	//		hr = pIDBProperties_->SetProperties(1, rgInitPropSet);
+	//		if (FAILED(hr))
+	//			throw std::runtime_error("Failed to set initialization properties.");
+
+	//		SysFreeString(InitProperties[0].vValue.bstrVal);
+	//		SysFreeString(InitProperties[1].vValue.bstrVal);
+	//		SysFreeString(InitProperties[2].vValue.bstrVal);
+	//		SysFreeString(InitProperties[3].vValue.bstrVal);
+	//		SysFreeString(InitProperties[4].vValue.bstrVal);
+
+	//		pIDBProperties_->Release();
+	//		pIDBProperties_ = nullptr;
+
+	//		// Now establish the connection to the data source.  
+	//		hr = pIDBInitialize_->Initialize();
+	//		if (FAILED(hr))
+	//			throw std::runtime_error("Failed to establish connection with the server.");
+	//	}
+	//	catch (...)
+	//	{
+	//		if (pIDBProperties_)
+	//		{
+	//			SysFreeString(InitProperties[0].vValue.bstrVal);
+	//			SysFreeString(InitProperties[1].vValue.bstrVal);
+	//			SysFreeString(InitProperties[2].vValue.bstrVal);
+	//			SysFreeString(InitProperties[3].vValue.bstrVal);
+	//			SysFreeString(InitProperties[4].vValue.bstrVal);
+
+	//			pIDBProperties_->Release();
+	//			pIDBProperties_ = nullptr;
+	//		}
+
+	//		if (FAILED(hr))
+	//		{
+	//			if (pIDBInitialize_)
+	//			{
+	//				pIDBInitialize_->Release();
+	//				pIDBInitialize_ = nullptr;
+	//			}
+	//		}
+
+	//		std::throw_with_nested(std::runtime_error("SQLConnection::initialize_and_establish_connection()"));
+	//	}
+
+	//	if (pIDBProperties_)
+	//	{
+	//		SysFreeString(InitProperties[0].vValue.bstrVal);
+	//		SysFreeString(InitProperties[1].vValue.bstrVal);
+	//		SysFreeString(InitProperties[2].vValue.bstrVal);
+	//		SysFreeString(InitProperties[3].vValue.bstrVal);
+	//		SysFreeString(InitProperties[4].vValue.bstrVal);
+	//	}
+
+	//	return hr;
 	//}
 
-	SQLConnection::SQLConnection(std::string db_data_source_name, std::string db_user_id, std::string db_user_password)
+	HRESULT SQLConnection::initialize_and_establish_connection(LPCWSTR lpwszProviderString)
 	{
-		hstmt1 = SQL_NULL_HSTMT;
-		hdbc1 = SQL_NULL_HDBC;
-		henv = SQL_NULL_HENV;
+		DBPROP          InitProperties[4] = { 0 };
+		DBPROPSET       rgInitPropSet[1] = { 0 };
+		HRESULT         hr = S_OK;
 
-		connect(db_data_source_name, db_user_id, db_user_password);
+		try
+		{
+			int attempts = 0;
+
+			do
+			{
+				attempts++;
+
+				// Obtain access to the OLE DB Driver for SQL Server.  
+				hr = CoCreateInstance(MSOLEDBSQL_CLSID,
+					NULL,
+					CLSCTX_INPROC_SERVER,
+					IID_IDBInitialize,
+					(void**)&pIDBInitialize_);
+
+				if (FAILED(hr))
+				{
+					char error_message[80];
+					sprintf_s(error_message, 80, "Attempt %d.  Failed to obtain access to the OLE DB Driver.  hr = %lX", attempts, hr);
+
+					std::cout << error_message << std::endl;
+
+					std::this_thread::sleep_for(std::chrono::milliseconds(500));
+
+					if (attempts == 3)
+						throw std::runtime_error(error_message);
+				}
+			} while (FAILED(hr));
+
+			// Initialize property values needed to establish connection.  
+			for (int i = 0; i < 4; i++)
+				VariantInit(&InitProperties[i].vValue);
+
+			// Server name.  
+			// See DBPROP structure for more information on InitProperties  
+			InitProperties[0].dwPropertyID = DBPROP_INIT_PROVIDERSTRING;
+			InitProperties[0].vValue.vt = VT_BSTR;
+			InitProperties[0].vValue.bstrVal = SysAllocString(lpwszProviderString);
+			InitProperties[0].dwOptions = DBPROPOPTIONS_REQUIRED;
+			InitProperties[0].colid = DB_NULLID;
+
+			// Construct the DBPROPSET structure(rgInitPropSet). The   
+			// DBPROPSET structure is used to pass an array of DBPROP   
+			// structures (InitProperties) to the SetProperties method.  
+			rgInitPropSet[0].guidPropertySet = DBPROPSET_DBINIT;
+			rgInitPropSet[0].cProperties = 1;
+			rgInitPropSet[0].rgProperties = InitProperties;
+
+			// Set initialization properties.  
+			hr = pIDBInitialize_->QueryInterface(IID_IDBProperties, (void**)&pIDBProperties_);
+			if (FAILED(hr))
+				throw std::runtime_error("Failed to obtain an IDBProperties interface.");
+
+			hr = pIDBProperties_->SetProperties(1, rgInitPropSet);
+			if (FAILED(hr))
+				throw std::runtime_error("Failed to set initialization properties.");
+
+			SysFreeString(InitProperties[0].vValue.bstrVal);
+
+			pIDBProperties_->Release();
+			pIDBProperties_ = nullptr;
+
+			// Now establish the connection to the data source.  
+			hr = pIDBInitialize_->Initialize();
+			if (FAILED(hr))
+				throw std::runtime_error("Failed to establish connection with the server.");
+		}
+		catch (...)
+		{
+			if (pIDBProperties_)
+			{
+				SysFreeString(InitProperties[0].vValue.bstrVal);
+				SysFreeString(InitProperties[1].vValue.bstrVal);
+				SysFreeString(InitProperties[2].vValue.bstrVal);
+				SysFreeString(InitProperties[3].vValue.bstrVal);
+
+				pIDBProperties_->Release();
+				pIDBProperties_ = nullptr;
+			}
+
+			if (FAILED(hr))
+			{
+				if (pIDBInitialize_)
+				{
+					pIDBInitialize_->Release();
+					pIDBInitialize_ = nullptr;
+				}
+			}
+
+			std::throw_with_nested(std::runtime_error("SQLConnection::initialize_and_establish_connection()"));
+		}
+
+		if (pIDBProperties_)
+		{
+			SysFreeString(InitProperties[0].vValue.bstrVal);
+			SysFreeString(InitProperties[1].vValue.bstrVal);
+			SysFreeString(InitProperties[2].vValue.bstrVal);
+			SysFreeString(InitProperties[3].vValue.bstrVal);
+		}
+
+		return hr;
+	}
+
+	SQLConnection::SQLConnection()
+	{
+		pIDBInitialize_ = NULL;
+	}
+
+	SQLConnection::SQLConnection(const std::string server, const std::string dbString, const std::string userID, const std::string password)
+	{
+		connect(server, dbString, userID, password);
 	}
 
 	SQLConnection::~SQLConnection()
 	{
-		cleanup();
+		disconnect();
 	}
 
-	void SQLConnection::cleanup()
+	void SQLConnection::connect(const std::string server, const std::string dbString, const std::string userID, const std::string password)
 	{
-		if (hstmt1 != SQL_NULL_HSTMT)
-			SQLFreeHandle(SQL_HANDLE_STMT, hstmt1);
+		HRESULT             hr = S_OK;
 
-		if (hdbc1 != SQL_NULL_HDBC) {
-			SQLDisconnect(hdbc1);
-			SQLFreeHandle(SQL_HANDLE_DBC, hdbc1);
-		}
+		// Initialize the Component Object Module Library
+		CoInitialize(nullptr);
 
-		if (henv != SQL_NULL_HENV)
-			SQLFreeHandle(SQL_HANDLE_ENV, henv);
-
-		hstmt1 = SQL_NULL_HSTMT;
-		hdbc1 = SQL_NULL_HDBC;
-		henv = SQL_NULL_HENV;
-	}
-
-	void SQLConnection::connect(const std::string data_source_name, const std::string userID, const std::string password)
-	{
-		RETCODE retcode;
-
-		// Allocate the ODBC environment and save handle.  
-		retcode = SQLAllocHandle(SQL_HANDLE_ENV, NULL, &henv);
-
-		if ((retcode != SQL_SUCCESS_WITH_INFO) && (retcode != SQL_SUCCESS))
+		try
 		{
-			char error_message[161];
-			SQLCHAR SqlState[6], Msg[SQL_MAX_MESSAGE_LENGTH];
-			SQLINTEGER NativeError;
-			SQLSMALLINT MsgLen;
+			//hr = initialize_and_establish_connection(Utilities::strtowstr(server).c_str(), 
+			//	Utilities::strtowstr(dbString).c_str(), 
+			//	Utilities::strtowstr(userID).c_str(), 
+			//	Utilities::strtowstr(password).c_str());
 
-			SQLGetDiagRec(SQL_HANDLE_DBC, hdbc1, 1, SqlState, &NativeError, Msg, sizeof(Msg), &MsgLen);
-			sprintf_s(error_message, 160, "Error allocating environment handle, retcode=%lX, SQLState=%s, NativeError=%d, Message=%s", retcode, SqlState, (int)NativeError, Msg);
+			//hr = initialize_and_establish_connection(L"Server=(local);Database=CleverLogsData;UID=TACollection-Ingestor;PWD=CD-Ing@TAColl!;Encrypt=Optional");
 
-			std::cout << error_message << std::endl;
 
-			cleanup();
+			//std::string input; // initialized elsewhere
+			//std::wstring output;
 
-			std::this_thread::sleep_for(std::chrono::milliseconds(500));
+			//output = std::wstring(input.begin(), input.end());
+			//output = output + std::wstring(L" program"); // or output += L" program";
+			//const wchar_t* ptr = output.c_str();
 
-			throw std::runtime_error(error_message);
+			//std::wstring = std::wstring(server.begin(), server.end());
+
+
+			//output = server << ";" << dbString << ";" << userID << ";" << password << ";Encrypt=Optional";
+
+			//hr = initialize_and_establish_connection(output.c_str());
+
+
+			//std::wstring output = Utilities::strtowstr(server)
+			//	+ Utilities::strtowstr(";").c_str()
+			//	+ Utilities::strtowstr(dbString).c_str()
+			//	+ Utilities::strtowstr(";").c_str()
+			//	+ Utilities::strtowstr(userID).c_str()
+			//	+ Utilities::strtowstr(";").c_str()
+			//	+ Utilities::strtowstr(password).c_str()
+			//	+ Utilities::strtowstr(";Encrypt=Optional").c_str();
+
+			//hr = initialize_and_establish_connection(output.c_str());
+
+			// Create the connection string
+			std::string connection_str = "Server=" + server + ";Database=" + dbString + ";UID=" + userID + ";PWD=" + password + ";Encrypt=Optional";
+
+			// newsize describes the length of the
+			// wchar_t string called wcstring in terms of the number
+			// of wide characters, not the number of bytes.
+			size_t newsize = strlen(connection_str.c_str()) + 1;
+
+			// The following creates a buffer large enough to contain
+			// the exact number of characters in the original string
+			// in the new format. If you want to add more characters
+			// to the end of the string, increase the value of newsize
+			// to increase the size of the buffer.
+			wchar_t* wcstring = new wchar_t[newsize];
+
+			// Convert char* string to a wchar_t* string.
+			size_t convertedChars = 0;
+			mbstowcs_s(&convertedChars, wcstring, newsize, connection_str.c_str(), _TRUNCATE);
+
+			// Establish connection to the SQL server.
+			hr = initialize_and_establish_connection(wcstring);
+
+			// Clean-up
+			delete[]wcstring;
+
+			if (FAILED(hr))
+				throw std::runtime_error("Failed to establish connection.");
 		}
-
-		// Set the ODBC version environment attribute
-		retcode = SQLSetEnvAttr(henv, SQL_ATTR_ODBC_VERSION, (SQLPOINTER)SQL_OV_ODBC3, SQL_IS_INTEGER);
-		if ((retcode != SQL_SUCCESS_WITH_INFO) && (retcode != SQL_SUCCESS))
+		catch (...)
 		{
-			char error_message[161];
-			SQLCHAR SqlState[6], Msg[SQL_MAX_MESSAGE_LENGTH];
-			SQLINTEGER NativeError;
-			SQLSMALLINT MsgLen;
+			if (pIDBInitialize_)
+			{
+				pIDBInitialize_->Uninitialize();
+				pIDBInitialize_->Release();
+				pIDBInitialize_ = nullptr;
+			}
 
-			SQLGetDiagRec(SQL_HANDLE_DBC, hdbc1, 1, SqlState, &NativeError, Msg, sizeof(Msg), &MsgLen);
-			sprintf_s(error_message, 160, "Error setting the ODBC version environment attribute, retcode=%lX, SQLState=%s, NativeError=%d, Message=%s", retcode, SqlState, (int)NativeError, Msg);
-
-			std::cout << error_message << std::endl;
-
-			cleanup();
-
-			std::this_thread::sleep_for(std::chrono::milliseconds(500));
-
-			throw std::runtime_error(error_message);
-		}
-
-		// Allocate a connection handle
-		SQLAllocHandle(SQL_HANDLE_DBC, henv, &hdbc1);
-		if ((retcode != SQL_SUCCESS_WITH_INFO) && (retcode != SQL_SUCCESS))
-		{
-			char error_message[161];
-			SQLCHAR SqlState[6], Msg[SQL_MAX_MESSAGE_LENGTH];
-			SQLINTEGER NativeError;
-			SQLSMALLINT MsgLen;
-
-			SQLGetDiagRec(SQL_HANDLE_DBC, hdbc1, 1, SqlState, &NativeError, Msg, sizeof(Msg), &MsgLen);
-			sprintf_s(error_message, 160, "Error allocating a connection handle, retcode=%lX, SQLState=%s, NativeError=%d, Message=%s", retcode, SqlState, (int)NativeError, Msg);
-
-			std::cout << error_message << std::endl;
-
-			cleanup();
-
-			std::this_thread::sleep_for(std::chrono::milliseconds(500));
-
-			throw std::runtime_error(error_message);
-		}
-
-		// Use SQL Server Authentication using a logon ID and password. Create the SQL Server DSN by using the SQL Server Authentication.
-		retcode = SQLConnect(hdbc1, (UCHAR*)data_source_name.c_str(), SQL_NTS, (UCHAR*)userID.c_str(), SQL_NTS, (UCHAR*)password.c_str(), SQL_NTS);
-		if ((retcode != SQL_SUCCESS_WITH_INFO) && (retcode != SQL_SUCCESS))
-		{
-			char error_message[161];
-			SQLCHAR SqlState[6], Msg[SQL_MAX_MESSAGE_LENGTH];
-			SQLINTEGER NativeError;
-			SQLSMALLINT MsgLen;
-
-			SQLGetDiagRec(SQL_HANDLE_DBC, hdbc1, 1, SqlState, &NativeError, Msg, sizeof(Msg), &MsgLen);
-			sprintf_s(error_message, 160, "Error connecting to the data source, retcode=%lX, SQLState=%s, NativeError=%d, Message=%s", retcode, SqlState, (int)NativeError, Msg);
-
-			std::cout << error_message << std::endl;
-
-			cleanup();
-
-			std::this_thread::sleep_for(std::chrono::milliseconds(500));
-
-			throw std::runtime_error(error_message);
-		}
-
-		// Allocate statement handle.  
-		retcode = SQLAllocHandle(SQL_HANDLE_STMT, hdbc1, &hstmt1);
-		if ((retcode != SQL_SUCCESS_WITH_INFO) && (retcode != SQL_SUCCESS))
-		{
-			char error_message[161];
-			SQLCHAR SqlState[6], Msg[SQL_MAX_MESSAGE_LENGTH];
-			SQLINTEGER NativeError;
-			SQLSMALLINT MsgLen;
-
-			SQLGetDiagRec(SQL_HANDLE_DBC, hdbc1, 1, SqlState, &NativeError, Msg, sizeof(Msg), &MsgLen);
-			sprintf_s(error_message, 160, "Error allocating statement handle, retcode=%lX, SQLState=%s, NativeError=%d, Message=%s", retcode, SqlState, (int)NativeError, Msg);
-
-			std::cout << error_message << std::endl;
-
-			cleanup();
-
-			std::this_thread::sleep_for(std::chrono::milliseconds(500));
-
-			throw std::runtime_error(error_message);
+			std::throw_with_nested(std::runtime_error("SQLConnection::initialize_and_establish_connection()"));
 		}
 	}
 
-	void SQLConnection::begin_transaction()
+	bool SQLConnection::connected()
 	{
-		RETCODE retcode;
-
-		// Begin the transaction
-		retcode = SQLSetConnectAttr(hdbc1, SQL_ATTR_AUTOCOMMIT, SQL_AUTOCOMMIT_OFF, SQL_IS_UINTEGER);
-		if ((retcode != SQL_SUCCESS) && (retcode != SQL_SUCCESS_WITH_INFO))
-		{
-			char error_message[161];
-			SQLCHAR SqlState[6], Msg[SQL_MAX_MESSAGE_LENGTH];
-			SQLINTEGER NativeError;
-			SQLSMALLINT MsgLen;
-
-			SQLGetDiagRec(SQL_HANDLE_DBC, hdbc1, 1, SqlState, &NativeError, Msg, sizeof(Msg), &MsgLen);
-			sprintf_s(error_message, 160, "Error starting the transaction, retcode=%lX, SQLState=%s, NativeError=%d, Message=%s", retcode, SqlState, (int)NativeError, Msg);
-
-			std::cout << error_message << std::endl;
-
-			throw std::runtime_error(error_message);
-		}
+		return pIDBInitialize_ != NULL;
 	}
 
-	void SQLConnection::commit_transaction()
+	void SQLConnection::disconnect()
 	{
-		SQLEndTran(SQL_HANDLE_DBC, hdbc1, SQL_COMMIT);
+		// Free Up All Allocated Memory
+		try
+		{
+			if (pIDBInitialize_)
+			{
+				pIDBInitialize_->Uninitialize();
+				pIDBInitialize_->Release();
+				pIDBInitialize_ = nullptr;
+			}
+
+			if (pIDBProperties_)
+			{
+				pIDBProperties_->Release();
+				pIDBProperties_ = nullptr;
+			}
+		}
+		catch (...)
+		{
+			std::ostringstream ss;
+			ss << ",method=SQLConnection.disconnect"
+				<< ",diagnostic_level=0"
+				<< ",exception=Unknown"
+				<< ",message=Error freeeing resources";
+			Utilities::logline_threadsafe << ss.str();
+
+			std::cerr << "SQLConnection::disconnect() - Error freeeing resources";
+		}
+
+		// Release The Component Object Module Library
+		CoUninitialize();
 	}
 }
