--- SOS/Domain/Develop Strategy V2/Run.DevelopStrategy_V2.h
@@ -1,8 +1,0 @@
-#pragma once
-
-namespace domain
-{
-	namespace develop_strategy_V2
-	{
-	}
-}
